% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_borough_features}
\alias{create_borough_features}
\title{Creates New York borough features for doing borough level map visualization}
\usage{
create_borough_features(zones_shapefile)
}
\arguments{
\item{zones_shapefile}{zones shapefile as an sf object, or downloads it from
https://s3.amazonaws.com/nyc-tlc/misc/taxi_zones.zip if it's left empty}
}
\value{
A \code{\link{SpatialPolygonsDataFrame}} with the borough's geometries and borough names
}
\description{
Reads zones shapefile provided by user, or download it from NYC website if not provided,
joins each borough's zones features together to create a geometry for each borough with its name attribute
}
\examples{
\dontrun{
   boroughs = create_borough_features()
}
}
\seealso{
\code{\link[nytaxi]{read_NYC_trip_dataset}} and \code{\link[nytaxi]{preprocess_dataset}}
}
