% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{download_NY_zones}
\alias{download_NY_zones}
\title{Download NYC zones from NYC website}
\usage{
download_NY_zones(as_sf = FALSE)
}
\arguments{
\item{as_sf}{if needed the zones can be returned as an sf object}
}
\value{
A \code{\link{SpatialPolygonsDataFrame}} with the NY zones
}
\description{
Downloads the zip file from NYC website (https://s3.amazonaws.com/nyc-tlc/misc/taxi_zones.zip)
unzips it to a temporary directory, read it and then delete the temporary files
}
\examples{
\dontrun{
   zones = download_NY_zones()
   #or
   zones = download_NY_zones(as_sf = T) #to download as sf object
}
}
