% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_boroughs_with_aggregate_trip_data}
\alias{get_boroughs_with_aggregate_trip_data}
\title{aggregates data from taxi trip data into the boroughs level}
\usage{
get_boroughs_with_aggregate_trip_data(taxidata, boroughs)
}
\arguments{
\item{taxidata}{The taxi trip data that were read from \code{\link[nytaxi]{read_NYC_trip_dataset}} and preprocessed \code{\link[nytaxi]{preprocess_dataset}}}

\item{boroughs}{is the New York zones sf object, if not present will be downloaded from NYC website}
}
\value{
A \code{\link{SpatialPolygonsDataFrame}} with the zones features with the \code{mean_tip_fair_ratio}, \code{mean_ratio_weekend} and \code{mean_ratio_weekday}
}
\description{
calculates mean tip-fare ratio for each borough for each day of the week, for each time of the day
and for weekend/non-weekend properties
}
\examples{
\dontrun{
   boroughs = create_borough_features()
   get_boroughs_with_aggregate_trip_data(taxidata, boroughs)
}
}
\seealso{
\code{\link[nytaxi]{read_NYC_trip_dataset}} and \code{\link[nytaxi]{preprocess_dataset}}
}
