% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_zones_with_aggregate_tripdata}
\alias{get_zones_with_aggregate_tripdata}
\title{Returns zones data along with aggregate taxi trip data}
\usage{
get_zones_with_aggregate_tripdata(taxidata, zones_newyork)
}
\arguments{
\item{taxidata}{The taxi trip data that were read from \code{\link[nytaxi]{read_NYC_trip_dataset}} and preprocessed \code{\link[nytaxi]{preprocess_dataset}}}

\item{zones_newyork}{is the New York zones sf object, if not present will be downloaded from NYC website}
}
\value{
A \code{\link{SpatialPolygonsDataFrame}} with the zones features with the \code{mean_tip_fair_ratio}, \code{mean_ratio_weekend} and \code{mean_ratio_weekday}
}
\description{
Reads the zones shapefile into SpatialPolygonsDataFrame, aggregates overall tip_fare ratio mean, and also for weekedays and weekends
and return back the SpatialPolygonsDataFrame
}
\examples{
\dontrun{
   url = "https://s3.amazonaws.com/nyc-tlc/trip+data/green_tripdata_2017-02.csv"
   taxidata = read_NYC_trip_dataset(url)
   get_zones_with_aggregate_tripdata(taxidata)
}
}
\seealso{
\code{\link[nytaxi]{read_NYC_trip_dataset}} and \code{\link[nytaxi]{preprocess_dataset}}
}
