% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preprocess_dataset}
\alias{preprocess_dataset}
\title{Preprocesses the New York taxi data CSV file to analyze}
\usage{
preprocess_dataset(taxidata, zones_newyork)
}
\arguments{
\item{taxidata}{is the dataset resulted from \code{\link[nytaxi]{read_NYC_trip_dataset}}}

\item{zones_newyork}{is the New York zones sf object, if not present will be downloaded from NYC website}
}
\value{
A \code{\link{data.frame}} that has the resulted dataset after preprocessing
}
\description{
Preprocessing includes checking the validity of the dataset, removing unneeded data columns and only leaves necesary ones,
parses data into the appropriate data types, creation of new columns that are necessary for analysis, adds and fills borough column for the
pickup and drop locations
}
\examples{
\dontrun{
   preprocess_dataset(read_NYC_trip_dataset())
}

}
\seealso{
\code{\link[nytaxi]{read_NYC_trip_dataset}}
}
