% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_NYC_trip_dataset}
\alias{read_NYC_trip_dataset}
\title{Reads the New Yrok taxi trip data}
\usage{
read_NYC_trip_dataset(url)
}
\arguments{
\item{url}{the url of the New York taxi data for analysis. Parameter is optional, if missing will use the example dataset of the library}
}
\value{
A \code{\link{data.frame}} with all the csv file for future preprocessing and processing
}
\description{
Reads the NYC Taxi & Limousine Commission green taxi trip  CSV file
}
\examples{
\dontrun{
   read_NYC_trip_dataset("https://s3.amazonaws.com/nyc-tlc/trip+data/green_tripdata_2017-02.csv")
   #Or
   read_NYC_trip_dataset("C:\\\\Downloads\\\\green_tripdata_2017-02.csv")
}
read_NYC_trip_dataset()
}
\seealso{
\code{\link[nytaxi]{preprocess_dataset}}
}
